+++ lib/hexapdf/type/acro_form/text_field.rb
@@ -254,9 +254,15 @@ module HexaPDF
         # :number::
         #     Assumes that the field value is a number and formats it according to the given
         #     arguments. See JavaScriptActions.af_number_format_action for details on the arguments.
+        #
+        # :percent::
+        #     Assumes that the field value is a number and formats it as percentage (where 1=100%
+        #     and 0=0%). See JavaScriptActions.af_percent_format_action for details on the
+        #     arguments.
         def set_format_action(type, **arguments)
           action_string = case type
                           when :number then JavaScriptActions.af_number_format_action(**arguments)
+                          when :percent then JavaScriptActions.af_percent_format_action(**arguments)
                           else
                             raise ArgumentError, "Invalid value for type argument: #{type.inspect}"
                           end